backend/requirements.txt
Flask
Flask-CORS
pymongo
gunicorn
python-dotenv
